2020.4.5 17 : 33 1.为新增类添加注释 2.修改LogController，在Cookie中添加ID 3.修改PlanController，添加在途订单拦截器 4.修改bug 5.修改车长和体积支持小数 5.添加对应JSP 6.修改对应的映射Mapper文件

+++ src/main/java/com/bishe/logistics_management/database/dao/CarDao.java
@@ -2,7 +2,6 @@ package com.bishe.logistics_management.database.dao;
 
 import com.bishe.logistics_management.database.dataObject.CarObject;
 import org.apache.ibatis.annotations.Mapper;
-import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Repository;
 
 import java.util.ArrayList;
@@ -52,7 +51,7 @@ public interface CarDao {
      * @param id 车辆ID
      * @param size 修改之后的载货空间
      */
-    void subSize(int id,int size);
+    void subSize(float id,float size);
 
     /**
      * 修改车辆的离港信息
